addressed comments from @yanavlasov
